Proof that clients support gzip and deflate compression
