+++ core/src/test/java/feign/client/DefaultClientTest.java
@@ -150,64 +150,6 @@ public class DefaultClientTest extends AbstractClientTest {
     assertThat(connection).isNotNull().isInstanceOf(HttpURLConnection.class);
   }
 
-
-  @Test
-  public void canSupportGzip() throws Exception {
-    /* enqueue a zipped response */
-    final String responseData = "Compressed Data";
-    server.enqueue(new MockResponse()
-        .addHeader("Content-Encoding", "gzip")
-        .setBody(new Buffer().write(compress(responseData))));
-
-    TestInterface api = newBuilder()
-        .target(TestInterface.class, "http://localhost:" + server.getPort());
-
-    String result = api.get();
-
-    /* verify that the response is unzipped */
-    assertThat(result).isNotNull()
-        .isEqualToIgnoringCase(responseData);
-
-  }
-
-  @Test
-  public void canExeptCaseInsensitiveHeader() throws Exception {
-    /* enqueue a zipped response */
-    final String responseData = "Compressed Data";
-    server.enqueue(new MockResponse()
-        .addHeader("content-encoding", "gzip")
-        .setBody(new Buffer().write(compress(responseData))));
-
-    TestInterface api = newBuilder()
-        .target(TestInterface.class, "http://localhost:" + server.getPort());
-
-    String result = api.get();
-
-    /* verify that the response is unzipped */
-    assertThat(result).isNotNull()
-        .isEqualToIgnoringCase(responseData);
-
-  }
-
-  @Test
-  public void canSupportDeflate() throws Exception {
-    /* enqueue a zipped response */
-    final String responseData = "Compressed Data";
-    server.enqueue(new MockResponse()
-        .addHeader("Content-Encoding", "deflate")
-        .setBody(new Buffer().write(deflate(responseData))));
-
-    TestInterface api = newBuilder()
-        .target(TestInterface.class, "http://localhost:" + server.getPort());
-
-    String result = api.get();
-
-    /* verify that the response is unzipped */
-    assertThat(result).isNotNull()
-        .isEqualToIgnoringCase(responseData);
-
-  }
-
   private byte[] compress(String data) throws Exception {
     try (ByteArrayOutputStream bos = new ByteArrayOutputStream(data.length())) {
       GZIPOutputStream gzipOutputStream = new GZIPOutputStream(bos);
