+++ googlehttpclient/README.md
@@ -2,7 +2,7 @@
 
 This module is a feign [Client](https://github.com/OpenFeign/feign/blob/master/core/src/main/java/feign/Client.java) to use the java [Google Http Client](https://github.com/googleapis/google-http-java-client).
 
-To use this, add to your classpath (via maven, or otherwise). Then cofigure Feign to use the GoogleHttpClient:
+To use this, add to your classpath (via maven, or otherwise). Then configure Feign to use the GoogleHttpClient:
 
 ```java
 GitHub github = Feign.builder()
