+++ googlehttpclient/src/test/java/feign/googlehttpclient/GoogleHttpClientTest.java
@@ -16,6 +16,7 @@ package feign.googlehttpclient;
 import feign.Feign;
 import feign.Feign.Builder;
 import feign.client.AbstractClientTest;
+import static org.junit.Assume.assumeFalse;
 
 public class GoogleHttpClientTest extends AbstractClientTest {
   @Override
@@ -34,4 +35,24 @@ public class GoogleHttpClientTest extends AbstractClientTest {
 
   @Override
   public void parsesUnauthorizedResponseBody() {}
+
+  /*
+   * Google HTTP client with NetHttpTransport does not support gzip and deflate compression
+   * out-of-the-box. You can replace the transport with Apache HTTP Client.
+   */
+  @Override
+  public void canSupportGzip() throws Exception {
+    assumeFalse("Google HTTP client client do not support gzip compression", false);
+  }
+
+  @Override
+  public void canSupportDeflate() throws Exception {
+    assumeFalse("Google HTTP client client do not support deflate compression", false);
+  }
+
+  @Override
+  public void canExceptCaseInsensitiveHeader() throws Exception {
+    assumeFalse("Google HTTP client client do not support gzip compression", false);
+  }
+
 }
